api/requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
tenacity
pytest
pytest-asyncio
